main: Listen to active window changed signal only when visible
We listen for the active window changed signal only to potentially
hide the manager window. So if we are hiding the window, then stop
listening to the signal. As the window can only be shown on explicit
"show" D-Bus method call, we start listening again to the signal when
we become visible.

The solution was stolen from eos-social.

This should fix the issue of the manager window not showing up on the
first press on the manager button, because we don't receive the signal
from Wnck when the focus jumps back and forth between the shell,
manager window and the last focused window.